pyproject.toml
[project]
name = "hackathon-climat-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "boto3>=1.42.0",
    "cartopy>=0.25.0",
    "fastparquet>=2024.11.0",
    "geopandas>=1.1.1",
    "fastexcel>=0.18.0",
    "h5netcdf>=1.7.3",
    "netcdf4>=1.7.3",
    "polars>=1.35.2",
    "requests>=2.32.5",
    "s3fs>=0.4.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "xarray>=2025.11.0",
    "dask>=2025.11.0",
    "pyarrow>=22.0.0",
    "cartiflette>=0.1.9",
    "fastapi>=0.123.5",
    "uvicorn[standard]>=0.38.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
]
